Refactor test login credentials
