+++ test/application_system_test_case.rb
@@ -1,6 +1,6 @@
 require "test_helper"
 
 class ApplicationSystemTestCase < ActionDispatch::SystemTestCase
-  include AuthHelper::Feature
+  include AuthForTest::Feature
   driven_by :selenium, using: :firefox, screen_size: [1400, 1400]
 end
